addressed original PR comments
