+++ src/workers/command.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use crate::{
     commands::*,
     consumers::{
@@ -9,6 +7,8 @@ use crate::{
     model::CapabilityConfig,
     APP_SPEC_ANNOTATION,
 };
+
+use super::insert_managed_annotations;
 
 /// A worker implementation for handling incoming commands
 #[derive(Clone)]
@@ -120,15 +120,4 @@ impl Worker for CommandWorker {
         }
         message.ack().await.map_err(WorkError::from)
     }
-}
-
-/// Inserts managed annotations to the given `annotations` HashMap.
-pub fn insert_managed_annotations(annotations: &mut HashMap<String, String>, model_name: &str) {
-    annotations.extend([
-        (
-            crate::MANAGED_BY_ANNOTATION.to_owned(),
-            crate::MANAGED_BY_IDENTIFIER.to_owned(),
-        ),
-        (APP_SPEC_ANNOTATION.to_owned(), model_name.to_owned()),
-    ])
 }
